Added config for site title, header and authorization block
The initializer can now accept values for site_title, which will
configure the page title.

The initializer can now accept values for header, which will configure
the branded header in the navigation.

The initializer can now accept an authorize block which is executed in
a controller context to detect administration privileges.