Changed the way authorization handled to allow session access
Authorization logic was happening scoped within the HelpKit module. This
has been changed and the #is_authorized method is now a helper method
that is also extended into the ApplicationController. This allows it to
be evaluated from within the controller instance. This will give that
method access to things such as the session object.